+$41 change REQ Thiết định thông báo

+++ app/src/main/java/jp/co/soliton/keymanager/SKMApplication.java
@@ -16,7 +16,7 @@ public class SKMApplication extends Application {
             CommonUtils.putPref(getApplicationContext(), StringList.KEY_OPENED_APP, new Boolean(true));
             CommonUtils.putPref(getApplicationContext(), StringList.KEY_NOTIF_ENABLE_FLAG, new Boolean(true));
             CommonUtils.putPref(getApplicationContext(), StringList.KEY_NOTIF_ENABLE_BEFORE_FLAG, new Boolean(true));
-            CommonUtils.putPref(getApplicationContext(), StringList.KEY_NOTIF_ENABLE_BEFORE, new Integer(7));
+            CommonUtils.putPref(getApplicationContext(), StringList.KEY_NOTIF_ENABLE_BEFORE, new Integer(14));
         }
     }
